Build docker

+++ ui/chains/cosmos/build.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 
-killall sifnoded sifnodecli
+if command -v killall &> /dev/null
+then
+  killall sifnoded sifnodecli
+fi
+
 
 rm $(which sifnoded) 2> /dev/null || echo sifnoded not install yet ...
 rm $(which sifnodecli) 2> /dev/null || echo sifnodecli not install yet ...
